AuditLogs implementation

+++ model/src/main/kotlin/org/katan/model/unit/auditlog/AuditLogEvent.kt
@@ -5,4 +5,4 @@ package org.katan.model.unit.auditlog
  *
  * @see AuditLogEntry
  */
-typealias AuditLogEvent = UByte
+typealias AuditLogEvent = UInt
